requirements.txt
Django>=3.2,<4.0
qrcode[pil]>=7.4.2
Pillow>=9.0.0
python-dateutil>=2.8.2
